affichage de chaque commande en detail sur administration

+++ inc/commandes/achat_gratuit.php
@@ -45,44 +45,69 @@ if (isset($_POST['achat_gratuit'])){
   }
 
   // ID ADR LIVR
-  // REQUETE POUR AVOIR L'ID DE L'ADRESSE DE LIVRAISON DU CLIENT
-  $req_adrlivr = $bdd->prepare("SELECT id_adresseclient FROM adresseclient WHERE id_client = ?");
+  // REQUETE POUR DEPLACER L'ADRESSE CLIENT AUX ARCHIVES
+  $req_adrlivr = $bdd->prepare("SELECT * FROM adresseclient WHERE id_client = ?");
   $req_adrlivr->execute(array($id_client));
   $result_livr = $req_adrlivr->fetch();
   //ID de l'adresse de livraison
-  $adr_livraison = $result_livr['id_adresseclient'];
+  $id_client = $result_livr['id_client'];
+  $Adresse= $result_livr['libelle_adresseclient'];
+  $comcomplet_livr= $result_livr['nomcomplet_adresseclient'];
+  $Societe= $result_livr['societe_adresseclient'];
+  $CP= $result_livr['codepostal_adresseclient'];
+  $etat_prov= $result_livr['etatprovince_adresseclient'];
+  $Ville= $result_livr['ville_adresseclient'];
+  $Pays= $result_livr['pays_adresseclient'];
+  $creationarchive = $bdd->prepare("INSERT INTO adressearchive(id_client,libelle_adresseclient,nomcomplet_adresseclient,societe_adresseclient,codepostal_adresseclient,etatprovince_adresseclient,ville_adresseclient,pays_adresseclient) VALUES(?, ?, ?, ?, ?, ?, ?, ?)");
+  $creationarchive->execute(array($id_client, $Adresse, $comcomplet_livr, $Societe, $CP, $etat_prov, $Ville, $Pays));
+  $result_archive = $creationarchive->rowCount();
 
-  //AJOUTER LA COMMANDE DANS LA TABLE COMMANDE SANS LE NUM COMMANDE ET FACTURE
-  $creationcommande = $bdd->prepare("INSERT INTO commande(datetime_commande, etat_commande, totalTTC, totalHT, fraisportHT, id_client, id_adresse_livraison, type_livraison, type_reglement, methode_reglement) VALUES(?, ?, ?, ?, ?, ?, ?, ?, ?, ?)");
-  $creationcommande->execute(array($date_commande, $etat_commande, $totalTTC, $totalHT, $fraisport, $id_client, $adr_livraison, $type_livraison, $type_regl, $methode_regl));
-  $result_com = $creationcommande->rowCount();
+  if($result_archive == 1){
 
-  if($result_com == 1){
+    //RECUP ID ADRESSE ARCHIVER
+    $adr_livraison = $bdd->lastInsertId();
 
-    $UID = $bdd -> lastInsertId(); //Recup ID de la commande
-    $numcommande = 'COM'.$UID;
-    $numfacture = 'FACT'.$UID;
+    //AJOUTER LA COMMANDE DANS LA TABLE COMMANDE SANS LE NUM COMMANDE ET FACTURE
+    $creationcommande = $bdd->prepare("INSERT INTO commande(datetime_commande, etat_commande, totalTTC, totalHT, fraisportHT, id_client, id_adresse_livraison, type_livraison, type_reglement, methode_reglement) VALUES(?, ?, ?, ?, ?, ?, ?, ?, ?, ?)");
+    $creationcommande->execute(array($date_commande, $etat_commande, $totalTTC, $totalHT, $fraisport, $id_client, $adr_livraison, $type_livraison, $type_regl, $methode_regl));
+    $result_com = $creationcommande->rowCount();
 
-    //Ajout des num fact et com
-    $ajoutfactcom = $bdd->prepare("UPDATE commande SET num_commande = ?, num_facture = ? WHERE id_commande = ? LIMIT 1");
-    $ajoutfactcom->execute(array($numcommande, $numfacture, $UID));
-    $result_factcom = $ajoutfactcom->rowCount();
+    if($result_com == 1){
 
-    if($result_factcom == 1){
-      //Suppression du panier qui vient d'être commandé
-      $deletecom = $bdd->prepare("DELETE FROM panier WHERE id_client = ? ");
-      $deletecom->execute(array($id_client));
-      header("Location: mercicommande.php");
+      $UID = $bdd -> lastInsertId(); //Recup ID de la commande
+      $numcommande = 'COM'.$UID;
+      $numfacture = 'FACT'.$UID;
+
+      //Ajout des num fact et com
+      $ajoutfactcom = $bdd->prepare("UPDATE commande SET num_commande = ?, num_facture = ? WHERE id_commande = ? LIMIT 1");
+      $ajoutfactcom->execute(array($numcommande, $numfacture, $UID));
+      $result_factcom = $ajoutfactcom->rowCount();
+
+      if($result_factcom == 1){
+        //Suppression du panier qui vient d'être commandé
+        $deletecom = $bdd->prepare("DELETE FROM panier WHERE id_client = ? ");
+        $deletecom->execute(array($id_client));
+        header("Location: mercicommande.php");
+      } else {
+        $errors[] = 'Problème lors du traitement de la commande !';
+        //Suppression de la commande qui n'a aucun num facture ni commande
+        $deletecom = $bdd->prepare("DELETE FROM commande WHERE id_commande = ? LIMIT 1");
+        $deletecom->execute(array($UID));
+        //Suppression de l'adresse archiver
+        $deletecom = $bdd->prepare("DELETE FROM adressearchive WHERE id_adresseclient = ? LIMIT 1");
+        $deletecom->execute(array($adr_livraison));
+        var_dump($ajoutfactcom->errorInfo());
+      }
     } else {
       $errors[] = 'Problème lors du traitement de la commande !';
-      //Suppression de la commande qui n'a aucun num facture ni commande
-      $deletecom = $bdd->prepare("DELETE FROM commande WHERE id_commande = ? LIMIT 1");
-      $deletecom->execute(array($UID));
-      var_dump($ajoutfactcom->errorInfo());
+      //Suppression de l'adresse archiver
+      $deletecom = $bdd->prepare("DELETE FROM adressearchive WHERE id_adresseclient = ? LIMIT 1");
+      $deletecom->execute(array($adr_livraison));
+      var_dump($creationcommande->errorInfo());
     }
   } else {
     $errors[] = 'Problème lors du traitement de la commande !';
-    var_dump($creationcommande->errorInfo());
+    var_dump($creationarchive->errorInfo());
   }
 }
 ?>
